+++ src-ui/src/app/components/filter-editor/filter-dropdown-date/filter-dropdown-date.component.scss
